implify a lot of things by only considering up to Integer.MAX_VALUE worth of meet when creating Concoctions; using more might well result in creating more items than KoL itself is willing to have inventory.
git-svn-id: http://svn.code.sf.net/p/kolmafia/code@19631 b29ace70-8910-0410-8dcc-aa2fc6433167

+++ src/net/sourceforge/kolmafia/request/CosmicRaysBazaarRequest.java
@@ -74,7 +74,7 @@ public class CosmicRaysBazaarRequest
 				return this.getCount() + " Meat";
 			}
 			@Override
-			public String getPluralName( long price )
+			public String getPluralName( int price )
 			{
 				return "Meat";
 			}
